Move TM2 combinedStates out of main manager
